fix tx_hash migration; only drop view if it exists

+++ lib/bitcoin/storage/sequel/migrations/005_change_tx_hash_to_bytea.rb
@@ -5,7 +5,7 @@ Sequel.migration do
     @log.info { "Running migration #{__FILE__}" }
 
     if adapter_scheme == :postgres
-      execute "DROP VIEW unconfirmed" # hasn't been used for a while now
+      execute "DROP VIEW unconfirmed"  if self.views.include?(:unconfirmed)
       execute "ALTER TABLE tx ALTER COLUMN hash TYPE bytea USING hash::bytea"
     end
